base work for new dashboard view
