+++ src/styles/nano/_cards.scss
@@ -10,7 +10,7 @@
 /**
 *   HOME CARDS
 **/
-.home-cards { background-color:lighten($grey, 2%); justify-content:space-between; min-height:$home-cards-min-height;
+.home-cards { background-color:rgba(lighten($grey, 2%), .5); justify-content:space-between; min-height:$home-cards-min-height;
     .mdl-card__supporting-text { width:auto; }
     .mdl-card__actions { text-align:right; }
     .mdl-card__media { background-color:transparent; }
@@ -35,3 +35,7 @@
     .completion-figure { margin:0; }
 }
 /*** specific home card | WEEKLY-CHART-CARD ***/
+
+/**
+*   CARDS STYLES GENERALS (V2)
+**/
